Move Mappable helpers from JSON to JSON+Mappable file
